Move generic logic from MetadataListPage to LinkedDataListPage
